Add CSVJSON.__init__(); extract dump_json()
This moves initial setup of the CSVJSON export process into an `__init__`
method to keep it separate from the actual conversion work of `main()`.

This moves some config, like `stream` and `json_kwargs` to be properties
of our `CSVJSON` object and thus requires raising the inner `dump_json()`
function to be an instance method.